PortAuthority now lives in neo4j-com FailSafe config not using parallel=classes as that somehow makes things fail
