+++ enterprise/causal-clustering/pom.xml
@@ -203,13 +203,6 @@
         </dependency>
 
         <dependency>
-            <groupId>org.neo4j.build</groupId>
-            <artifactId>port-authority</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
             <scope>test</scope>
@@ -262,7 +255,6 @@
                 <configuration>
                     <!-- Integration tests spend significant time waiting, so we can run more things in parallel here-->
                     <forkCount>2C</forkCount>
-                    <parallel>classes</parallel>
                     <!-- Misha says it is not safe to run tests with more than one thread per JVM -->
                     <!-- He also said he'd fix that! -->
                     <threadCount>1</threadCount>
