+++ enterprise/com/src/test/java/org/neo4j/com/ports/allocation/DefaultPortProbe.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.build.portauthority;
+package org.neo4j.com.ports.allocation;
 
 import java.io.IOException;
 import java.net.ServerSocket;
