+++ enterprise/com/src/test/java/org/neo4j/com/ports/allocation/PortRepositoryIT.java
@@ -17,22 +17,18 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.build.portauthority;
+package org.neo4j.com.ports.allocation;
 
 import static java.util.Arrays.asList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
-import static org.neo4j.build.portauthority.PortConstants.EphemeralPortMinimum;
+import static org.neo4j.com.ports.allocation.PortConstants.EphemeralPortMinimum;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.HashSet;
 
-import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
