+++ enterprise/ha/pom.xml
@@ -50,17 +50,6 @@
     </license>
   </licenses>
 
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-failsafe-plugin</artifactId>
-        <configuration>
-          <forkCount>1</forkCount>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-
   <dependencies>
     <dependency>
       <groupId>org.neo4j</groupId>
@@ -215,4 +204,29 @@
       <version>${project.version}</version>
     </dependency>
   </dependencies>
+
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-failsafe-plugin</artifactId>
+        <configuration>
+          <!-- Integration tests spend significant time waiting, so we can run more things in parallel here-->
+          <forkCount>2C</forkCount>
+          <!-- Misha says it is not safe to run tests with more than one thread per JVM -->
+          <!-- He also said he'd fix that! -->
+          <threadCount>1</threadCount>
+          <reuseForks>true</reuseForks>
+          <runOrder>random</runOrder>
+          <!-- slowest test class takes up to 10 minutes, so let's cater for that, with contingency -->
+          <parallelTestsTimeoutInSeconds>1200</parallelTestsTimeoutInSeconds>
+          <parallelTestsTimeoutForcedInSeconds>1800</parallelTestsTimeoutForcedInSeconds>
+          <workingDirectory>${project.build.directory}/FORK_DIRECTORY_${surefire.forkNumber}</workingDirectory>
+          <systemPropertyVariables>
+            <port.authority.directory>${user.dir}/target/port-authority-${maven.build.timestamp}</port.authority.directory>
+          </systemPropertyVariables>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
 </project>
