+++ enterprise/ha/src/test/java/org/neo4j/kernel/ha/ConflictingServerIdIT.java
@@ -24,8 +24,8 @@ import org.junit.Test;
 
 import java.io.File;
 
-import org.neo4j.build.portauthority.PortAuthority;
 import org.neo4j.cluster.ClusterSettings;
+import org.neo4j.com.ports.allocation.PortAuthority;
 import org.neo4j.graphdb.Transaction;
 import org.neo4j.graphdb.factory.GraphDatabaseBuilder;
 import org.neo4j.graphdb.factory.TestHighlyAvailableGraphDatabaseFactory;
